// src/app/models/standingsArray.interface.ts
// tslint:disable-next-line: class-name
export interface StandingsArray {
        Conference: string;
        Division: string;
        Wins: number;
        Losses: number;
        NetPoints: number;
        Touchdowns: number;
        DivisionWins: number;
        DivisionLosses: number;
        ConferenceWins: number;
        ConferenceLosses: number;
        GlobalTeamID: number;
        DivisionRank: number;
        ConferenceRank: number;
}